requirements.txt
streamlit==1.52.1
pandas==2.3.3
numpy==1.26.4
python-dotenv==1.0.0

openai==2.9.0
exa-py==2.0.1

SQLAlchemy==2.0.44
pyarrow==22.0.0
lancedb==0.4.1

sentence-transformers==2.2.2
torch==2.2.2
transformers==4.33.3

google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-api-python-client==2.187.0

requests==2.32.5
tqdm==4.67.1
protobuf==6.33.2
